(1) Term "Property" is now "Feature" with respect to statistical feature (mean, median, ...) (2) Simplified Aggregator interface (3) Moved WeightFn to upper level

+++ calvalus-processing/src/main/java/com/bc/calvalus/processing/l3/L3Config.java
@@ -70,7 +70,7 @@ public class L3Config {
 
         Double weightCoeff;
 
-        Double fillValue;
+        Float fillValue;
 
         public AggregatorConfiguration() {
         }
@@ -99,7 +99,7 @@ public class L3Config {
             return weightCoeff;
         }
 
-        public Double getFillValue() {
+        public Float getFillValue() {
             return fillValue;
 
         }
@@ -124,7 +124,7 @@ public class L3Config {
             this.weightCoeff = weightCoeff;
         }
 
-        public void setFillValue(Double fillValue) {
+        public void setFillValue(Float fillValue) {
             this.fillValue = fillValue;
         }
     }
